FIxed introduced regression

+++ app/models/users/photos.model.js
@@ -110,6 +110,9 @@ exports.deletePhoto = async (token, id, done) => {
       }
 
       const { filename } = rows[0][0];
+      if (!isStringAndNotEmpty(filename)) {
+        return done(404);
+      }
 
       if (fs.existsSync(filename)) {
         fs.unlinkSync(filename);
